// applications/panel-webapp/src/services/openapi/types.gen.ts
export type ProductSchema = {
    /**
     * Product Code
     */
    product_code: string;
    /**
     * Product Name
     */
    product_name: string;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Price
     */
    price: number;
    /**
     * Quantity
     */
    quantity: number;
    /**
     * Category
     */
    category?: string | null;
    /**
     * Unit
     */
    unit?: string | null;
    /**
     * Shelf
     */
    shelf?: string | null;
    /**
     * Image Id
     */
    image_id?: Array<string>;
};